refactor code related to header
