--- web/src/components/header.module.css
@@ -1,102 +1,0 @@
-@import url("../styles/index.css");
-
-header {
-    border-bottom: 5px solid;
-    display: flex;
-    justify-content: center;
-    background-color: white;
-    /* position relative below for dropdown to allign with the with */
-    position: relative;
-}
-.header_container {
-    display: flex;
-    padding: 0 20px;
-    width: 1100px;
-
-    /* height: 80px; */
-}
-
-/* .header_border_blue {
-    border-bottom-color: var(--blue);
-}
-
-.header_border_green {
-    border-bottom-color: var(--green);
-}
-
-.header_border_orange {
-    border-bottom-color: var(--orange);
-} */
-
-.header_border_red {
-    border-bottom-color: var(--red);
-}
-
-.rightside_wrapper {
-    /* width: 100px; */
-    display: flex;
-    flex-direction: column;
-
-    justify-content: flex-end;
-    /* margin-left: 10px; */
-}
-
-.rightside_wrapper > a > img {
-    height: 30px;
-}
-
-@media (max-width: 991px) {
-    /* .header_container {
-        padding: 10px;
-    } */
-    /* .logo_wrapper {
-        width: 150px;
-    } */
-    /* img {
-        height: 70px;
-    } */
-    .arrow {
-        padding:5px;
-    }
-    .arrow_up {
-        top: 5px;
-    }
-}
-
-@media (max-width: 767px) {
-    header {
-        height: 70px;
-        border: none;
-    }
-    .header_container {
-        width: 100%;
-        justify-content: flex-end;
-        padding: 0 5px;
-        align-items: center;
-        position: absolute;
-        top: 0;
-        left: 0;
-        background-color: white;
-        border-bottom: 5px solid red;
-        z-index: 999;
-    }
-    .logo_wrapper {
-        height: 70px;
-        flex-grow: 1;
-        display: flex;
-        justify-content: center;
-    }
-    .logo {
-        /* align-self: center; */
-        height: 70px;
-        /* flex-grow: 1; */
-    }
-    .rightside_wrapper {
-        flex-direction: row;
-        margin-left: 0px;
-        width: 54px;
-    }
-    .open_header {
-        /* position: fixed; */
-    }
-  }
