+++ web/src/components/header/components/dropdownParent.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import styles from "./navbar.module.css";
+import styles from "../styles/header.module.css";
 
 const DropdownParent = ({ title, children }) => {
   const [openDropdown, setOpenDropdown] = useState(false);
