+++ web/src/components/header/data/dropdownItemsQuery.js
