--- web/src/components/news/news-post-preview-grid.js
@@ -1,32 +1,0 @@
-// import { Link } from 'gatsby'
-// import React from 'react'
-// import NewsPostPreview from './news-post-preview'
-
-// function NewsPostPreviewGrid(props) {
-//   return (
-//     <div>
-//       {props.title && <h2>{props.title}</h2>}
-//       <ul>
-//         {props.nodes &&
-//           props.nodes.map(node => (
-//             <li key={node.id}>
-//               <NewsPostPreview {...node} />
-//             </li>
-//           ))}
-//       </ul>
-//       {props.browseMoreHref && (
-//         <div>
-//           <Link to={props.browseMoreHref}>Browse more</Link>
-//         </div>
-//       )}
-//     </div>
-//   )
-// }
-
-// NewsPostPreviewGrid.defaultProps = {
-//   title: '',
-//   nodes: [],
-//   browseMoreHref: ''
-// }
-
-// export default NewsPostPreviewGrid
