--- web/src/components/toggler.module.css
@@ -1,50 +1,0 @@
-@import url("../styles/index.css");
-
-.button {
-    height: 44px;
-    width: 44px;
-    margin: 5px;
-    background: var(--gray);
-    border: none;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-    align-items: center;
-    border: 2px solid white;
-    z-index: 2;
-}
-.button:focus {
-    outline: none;
-}
-
-.line {
-    width: 30px;
-    height: 3px;
-    background: white;
-    transition: all 0.4s;
-}
-
-.line1 {
-    transform: rotate(-45deg) translate(-8px, 8px);
-  }
-
-  .line2 {
-    opacity: 0;
-  }
-
-  .line3 {
-    transform: rotate(45deg) translate(-7px, -7px);
-  }
-
-  /* @media (min-width: 768px) {
-    .button {
-        display: none;
-    }
-  }
-  @media (max-width: 767px) {
-    .button {
-        position: fixed;
-        top: 5px;
-        left: 20px
-    }
-  } */
